+++ client45/src/main/java/businesslogic/receiptbl/ReceiptTrans.java
@@ -231,7 +231,7 @@ public class ReceiptTrans {
 		String id = po.getID();
 		ReceiptType type = po.getReceiptType();
 		String courierID = po.getCourierID();
-		double money = po.getMoney();
+		BigDecimal money = po.getMoney();
 		ArrayList<String> orderNumbers = po.getOrderNumbers();
 		String date = po.getDate();
 		return new DebitBillVO(id, type, courierID, money, orderNumbers, date);
@@ -240,7 +240,7 @@ public class ReceiptTrans {
 		String id = vo.ID;
 		ReceiptType type = vo.type;
 		String courierID =vo.courierID;
-		double money = vo.money;
+		BigDecimal money = vo.money;
 		ArrayList<String> orderNumbers = vo.orderNumbers;
 		String date = vo.date;
 		return new DebitBillPO(id, type, courierID, money, orderNumbers, date);
@@ -249,7 +249,7 @@ public class ReceiptTrans {
 		String iD = po.getID();
 		String date = po.getDate();
 		ReceiptType type = po.getReceiptType();
-		double money = po.getMoney();
+		BigDecimal money = po.getMoney();
 		String payerName = po.getPayerName();
 		String accountID = po.getAccountID();
 		PayBillItem items = po.getPayBillItem();
@@ -260,7 +260,7 @@ public class ReceiptTrans {
 		String ID = vo.ID;
 		String date = vo.date;
 		ReceiptType type = vo.type;
-		double money =vo.money;
+		BigDecimal money =vo.money;
 		String payerName = vo.payerName;
 		String accountID = vo.accountID;
 		PayBillItem items = vo.items;
